mb/system76/adl: darp8,lemp11: Disable RTD3 on SATA port
After switching to S3, it was found that drives on the SATA port do not
exit D3cold on S3 exit. Disable RTD3 on the port until the issue can be
resolved.

Avoids the following error in Linux:

    pcieport 0000:00:1d.0: Unable to change power state from D3cold to D0, device inaccessible

Tested on darp8 with a Samsung 970 EVO or Crucial P5 in J_SSD1.

Change-Id: Ib26f59db61acfbf9248cea379c197765d3d9c470
Signed-off-by: Tim Crawford <tcrawford@system76.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/76593
Reviewed-by: Jeremy Soller <jeremy@system76.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
(cherry picked from commit e56c738f323c72574444152dc5365a8feec52118)